review: move Copyable to file CopyButton.tsx
